requirements.txt
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.18
botocore==1.35.18
certifi==2024.7.4
charset-normalizer==3.4.0
Django==5.1
django-cors-headers==4.4.0
django-storages==1.14.4
django4-background-tasks==1.2.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.3.4
drf-yasg==1.21.7
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.1
s3transfer==0.10.2
sendgrid==6.11.0
six==1.16.0
sqlparse==0.5.1
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
